feat: 조회,추가,수정,삭제 기능 구현

+++ src/main/java/gift/Product.java
@@ -1,25 +1,25 @@
 package gift;
 
 public class Product {
-    private int id;
+    private long id;
     private String name;
     private int price;
     private String imageUrl;
 
     public Product(){}
 
-    public Product(int id, String name, int price, String imageUrl) {
+    public Product(long id, String name, int price, String imageUrl) {
         this.id = id;
         this.name = name;
         this.price = price;
         this.imageUrl = imageUrl;
     }
 
-    public int getId(){
+    public long getId(){
         return id;
     }
 
-    public void setId(int id){
+    public void setId(long id){
         this.id = id;
     }
